+++ PrismaticSlime/docs/Changelog.md
@@ -4,7 +4,8 @@
 <!---
 TODO:
 * Make new buff
-* Implement random buff for the prismatic toast...
+* Fix up buff indicator for the prismatic jelly toast.
+* Figure out how to get the player to glow different colors.
 * Figure out dye pots and make them free if given a prismatic jelly. Add Emily event.
 * Wizard gift + payment
 * Everyone else gets 75 heart points? -->